+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Changed
+
+- Replace conversion/transfer type wording to be `join a MAT`.
+
 ### Fixed
 
 - Clear any radio button values in the Sponsored support grant type task if the